pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seugearbox"
version = "0.1.0"
description = "Reproducible gearbox fault diagnosis experiments with HHO-BP."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "Pillow>=10.0",
  "pyts>=0.13",
  "PyWavelets>=1.6",
  "scikit-learn>=1.4",
  "scipy>=1.12",
  "seaborn>=0.13",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["seugearbox*"]
